requirements.txt

playwright>=1.40.0
pytest==8.3.2
pytest-xdist==3.6.1
PyYAML==6.0.2
lingua-language-detector==2.1.1
pillow==10.4.0
babel==2.15.0
rapidfuzz==3.9.6
pandas==2.1.4
openpyxl==3.1.2
numpy==1.24.3
